// src/app/modules/icms/ifb/models/ifb.ts
export class IFB {
    id: number;
    ifbDate: Date; 
    caseId: string; 
    accountNumber: string; 
    borrowerName: string; 
    productTypes: string; 
    otherProductTypes: string; 
    authorizedBy: string; 
    authorizationTimeStamp: string; 
    allCategory: any; 
    allSubCategory: any; 
    irregularity: string; 
    amountInvolved: string; 
    responsiblePerson: string; 
    status: any; 
}